+++ docs/assets/navigation.js
@@ -1,0 +1,1 @@
+window.navigationData = "eJyLrlYqSa0oUbJSSklNSyzNKVHSUSpILMlQslIqSyzKTEzKSS3Wh0rpZZTk5ijpKGVn5qUoWRkb1cYCAKAYFR0="