+++ CHANGELOG.md
@@ -94,7 +94,7 @@ All notable changes to this project will be documented in this file. This projec
 ## [4.1.0](https://github.com/OpenWonderLabs/homebridge-switchbot/releases/tag/v4.1.0) (2024-10-21)
 
 ### What's Changed
-- Add `Meter Pro` deviceType Support
+- Add `MeterPro` deviceType Support
 - Fix `node-switchbot` noble issue.
 - Housekeeping and updated dependencies.